# Performance Metrics API: Refactored the AZ EventLogger interface