--- a/Code/Framework/AzCore/AzCore/IO/TextStreamWriters.h
+++ b/Code/Framework/AzCore/AzCore/IO/TextStreamWriters.h
@@ -5,184 +5,193 @@
  * SPDX-License-Identifier: Apache-2.0 OR MIT
  *
  */
-#ifndef AZCORE_IO_TEXTSTREAMWRITERS_H
-#define AZCORE_IO_TEXTSTREAMWRITERS_H
+#pragma once
 
 #include <AzCore/base.h>
 #include <AzCore/IO/GenericStreams.h>
 #include <AzCore/std/containers/vector.h>
 
-namespace AZ
+namespace AZ::IO
 {
-    namespace IO
+    /**
+     * For use with rapidxml::print.
+     */
+    class RapidXMLStreamWriter
     {
-        /**
-         * For use with rapidxml::print.
-         */
-        class RapidXMLStreamWriter
+    public:
+        class PrintIterator
         {
         public:
-            class PrintIterator
+            PrintIterator(RapidXMLStreamWriter* writer)
+                : m_writer(writer)
             {
-            public:
-                PrintIterator(RapidXMLStreamWriter* writer)
-                    : m_writer(writer)
-                {
-                }
-
-                PrintIterator(const PrintIterator& it)
-                    : m_writer(it.m_writer)
-                {
-                }
-
-                PrintIterator& operator=(const PrintIterator& rhs) = default;
-
-                PrintIterator& operator*() { return *this; }
-                PrintIterator& operator++() { return *this; }
-                PrintIterator& operator++(int) { return *this; }
-                PrintIterator& operator=(char c)
-                {
-                    if (m_writer->m_cache.size() == m_writer->m_cache.capacity())
-                    {
-                        m_writer->FlushCache();
-                    }
-                    m_writer->m_cache.push_back(c);
-                    return *this;
-                }
-            private:
-                RapidXMLStreamWriter* m_writer;
-            };
-
-            RapidXMLStreamWriter(IO::GenericStream* stream, size_t writeCacheSize = 128 * 1024)
-                : m_stream(stream)
-                , m_errorCount(0) 
-            {
-                m_cache.reserve(writeCacheSize);
             }
 
-            ~RapidXMLStreamWriter()
+            PrintIterator(const PrintIterator& it)
+                : m_writer(it.m_writer)
             {
-                FlushCache();
             }
 
-            RapidXMLStreamWriter(const RapidXMLStreamWriter&) = delete;
-            RapidXMLStreamWriter& operator=(const RapidXMLStreamWriter&) = delete;
+            PrintIterator& operator=(const PrintIterator& rhs) = default;
 
-            void FlushCache()
+            PrintIterator& operator*() { return *this; }
+            PrintIterator& operator++() { return *this; }
+            PrintIterator& operator++(int) { return *this; }
+            PrintIterator& operator=(char c)
             {
-                if (!m_cache.empty())
+                if (m_writer->m_cache.size() == m_writer->m_cache.capacity())
                 {
-                    IO::SizeType bytes = m_stream->Write(m_cache.size(), m_cache.data());
-                    if (bytes != m_cache.size())
-                    {
-                        AZ_Error("Serializer", false, "Failed writing %d byte(s) to stream, wrote only %d bytes!", m_cache.size(), bytes);
-                        ++m_errorCount;
-                    }
-                    m_cache.clear();
+                    m_writer->FlushCache();
                 }
+                m_writer->m_cache.push_back(c);
+                return *this;
             }
-
-            PrintIterator Iterator()
-            {
-                return PrintIterator(this);
-            }
-
-            IO::GenericStream* m_stream;
-            AZStd::vector<AZ::u8> m_cache;
-            int m_errorCount;
+        private:
+            RapidXMLStreamWriter* m_writer;
         };
 
-        /**
-         * For use with rapidxml::Writer/PrettyWriter.
-         */
-        class RapidJSONStreamWriter
+        RapidXMLStreamWriter(IO::GenericStream* stream, size_t writeCacheSize = 128 * 1024)
+            : m_stream(stream)
+            , m_errorCount(0)
         {
-        public:
-            typedef char Ch;    //!< Character type. Only support char.
+            m_cache.reserve(writeCacheSize);
+        }
 
-            RapidJSONStreamWriter(AZ::IO::GenericStream* stream, size_t writeCacheSize = 128 * 1024)
-                : m_stream(stream)
+        ~RapidXMLStreamWriter()
+        {
+            FlushCache();
+        }
+
+        RapidXMLStreamWriter(const RapidXMLStreamWriter&) = delete;
+        RapidXMLStreamWriter& operator=(const RapidXMLStreamWriter&) = delete;
+
+        void FlushCache()
+        {
+            if (!m_cache.empty())
             {
-                m_cache.reserve(writeCacheSize);
+                IO::SizeType bytes = m_stream->Write(m_cache.size(), m_cache.data());
+                if (bytes != m_cache.size())
+                {
+                    AZ_Error("Serializer", false, "Failed writing %d byte(s) to stream, wrote only %d bytes!", m_cache.size(), bytes);
+                    ++m_errorCount;
+                }
+                m_cache.clear();
             }
+        }
 
-            ~RapidJSONStreamWriter()
+        PrintIterator Iterator()
+        {
+            return PrintIterator(this);
+        }
+
+        IO::GenericStream* m_stream;
+        AZStd::vector<AZ::u8> m_cache;
+        int m_errorCount;
+    };
+
+    //! Implements the rapidjson::Stream concept
+    class RapidJSONWriteStreamNoCache
+    {
+    public:
+        using Ch = char;    //!< Character type. Only support char.
+
+        explicit RapidJSONWriteStreamNoCache(AZ::IO::GenericStream& stream)
+            : m_stream(&stream)
+        {
+        }
+
+        virtual ~RapidJSONWriteStreamNoCache() = default;
+
+        RapidJSONWriteStreamNoCache(const RapidJSONWriteStreamNoCache&) = delete;
+        RapidJSONWriteStreamNoCache& operator=(const RapidJSONWriteStreamNoCache&) = delete;
+
+        virtual void Put(char c)
+        {
+            if (IO::SizeType bytes = m_stream->Write(1, &c); bytes != 1)
+            {
+                AZ_Error("Serializer", false, "Failed to add byte to stream, stream current size is %zu", m_stream->GetLength());
+            }
+        }
+
+        virtual void Flush()
+        {
+            // GenericStream doesn't support a flush API
+        }
+
+        // Not implemented
+        char Peek() const
+        {
+            AZ_Assert(false, "RapidJSONStreamWriter Peek not supported.");
+            return 0;
+        }
+
+        char Take()
+        {
+            AZ_Assert(false, "RapidJSONStreamWriter Take not supported.");
+            return 0;
+        }
+        size_t Tell() const
+        {
+            AZ_Assert(false, "RapidJSONStreamWriter Tell not supported.");
+            return 0;
+        }
+        char* PutBegin()
+        {
+            AZ_Assert(false, "RapidJSONStreamWriter PutBegin not supported.");
+            return 0;
+        }
+        size_t PutEnd(char*)
+        {
+            AZ_Assert(false, "RapidJSONStreamWriter PutEnd not supported.");
+            return 0;
+        }
+
+        AZ::IO::GenericStream* m_stream;
+    };
+
+    //! Adds caching around the RapidJsonStreamWriter
+    class RapidJSONStreamWriter
+        : public RapidJSONWriteStreamNoCache
+    {
+    public:
+        explicit RapidJSONStreamWriter(AZ::IO::GenericStream* stream, size_t writeCacheSize = 128 * 1024)
+            : RapidJSONWriteStreamNoCache(*stream)
+        {
+            m_cache.reserve(writeCacheSize);
+        }
+
+        ~RapidJSONStreamWriter()
+        {
+            FlushCache();
+        }
+
+        void Put(char c) override
+        {
+            if (m_cache.size() == m_cache.capacity())
             {
                 FlushCache();
             }
+            m_cache.push_back(c);
+        }
 
-            RapidJSONStreamWriter(const RapidJSONStreamWriter&) = delete;
-            RapidJSONStreamWriter& operator=(const RapidJSONStreamWriter&) = delete;
+        void Flush() override
+        {
+            FlushCache();
+        }
 
-            void FlushCache()
+    private:
+        void FlushCache()
+        {
+            if (!m_cache.empty())
             {
-                if (!m_cache.empty())
+                IO::SizeType bytes = m_stream->Write(m_cache.size(), m_cache.data());
+                if (bytes != m_cache.size())
                 {
-                    IO::SizeType bytes = m_stream->Write(m_cache.size(), m_cache.data());
-                    if (bytes != m_cache.size())
-                    {
-                        AZ_Error("Serializer", false, "Failed writing %d byte(s) to stream, wrote only %d bytes!", m_cache.size(), bytes);
-                    }
-                    m_cache.clear();
+                    AZ_Error("Serializer", false, "Failed writing %d byte(s) to stream, wrote only %d bytes!", m_cache.size(), bytes);
                 }
+                m_cache.clear();
             }
-
-            void Put(char c)
-            {
-                if (m_cache.size() == m_cache.capacity())
-                {
-                    FlushCache();
-                }
-                m_cache.push_back(c);
-            }
-
-            void PutN(char c, size_t n)
-            {
-                while (n > 0)
-                {
-                    Put(c); // if this is common we can optimize the write to the cache
-                    --n;
-                }
-            }
-
-            void Flush()
-            {
-                FlushCache();
-            }
-
-            // Not implemented
-            char Peek() const
-            {
-                AZ_Assert(false, "RapidJSONStreamWriter Peek not supported.");
-                return 0;
-            }
-
-            char Take()
-            {
-                AZ_Assert(false, "RapidJSONStreamWriter Take not supported.");
-                return 0;
-            }
-            size_t Tell() const
-            {
-                AZ_Assert(false, "RapidJSONStreamWriter Tell not supported.");
-                return 0;
-            }
-            char* PutBegin()
-            {
-                AZ_Assert(false, "RapidJSONStreamWriter PutBegin not supported.");
-                return 0;
-            }
-            size_t PutEnd(char*)
-            {
-                AZ_Assert(false, "RapidJSONStreamWriter PutEnd not supported.");
-                return 0;
-            }
-
-            AZ::IO::GenericStream* m_stream;
-            AZStd::vector<AZ::u8> m_cache;
-        };
-    }   // namespace IO
-}   // namespace AZ
-
-#endif  // AZCORE_IO_TEXTSTREAMWRITERS_H
-#pragma once
+        }
+        AZStd::vector<AZ::u8> m_cache;
+    };
+}   // namespace AZ::IO