--- a/Code/Framework/AzCore/AzCore/Metrics/IEventLoggerFactory.h
+++ b/Code/Framework/AzCore/AzCore/Metrics/IEventLoggerFactory.h
@@ -65,5 +65,4 @@ namespace AZ::Metrics
 
     using EventLoggerFactory = AZ::Interface<IEventLoggerFactory>;
 
-} // namespace AZ::Event::Metrics
-
+} // namespace AZ::Metrics