… Code/Framework/AzCore/AzCore/Debug/LocalFileEventLogger.cpp
… Code/Framework/AzCore/AzCore/Debug/LocalFileEventLogger.h
… Code/Framework/AzCore/AzCore/JSON/RapidjsonAllocatorAdapter.h
… Code/Framework/AzCore/AzCore/Metrics/IEventLogger.cpp
… Code/Framework/AzCore/AzCore/Metrics/IEventLogger.inl
… Code/Framework/AzCore/AzCore/Metrics/JsonTraceEventLogger.cpp
… Code/Framework/AzCore/AzCore/Metrics/JsonTraceEventLogger.h
… Code/Framework/AzCore/AzCore/azcore_files.cmake
… Code/Framework/AzCore/Tests/Debug/LocalFileEventLoggerTests.cpp
… Code/Framework/AzCore/Tests/Metrics/EventLoggerFactoryTests.cpp
… Code/Framework/AzCore/Tests/Metrics/JsonTraceEventLoggerTests.cpp
… Code/Framework/AzCore/Tests/azcoretests_files.cmake