## What does this PR do?
The EventLogger interface supports logging events in the google trace event format
(https://docs.google.com/document/d/1CvAClvFfyA5R-PhYUmn5OOQtYMH4h6I0nSsKchNAySU/)

The following events are supported: Duration Events, Complete Events, Instant Events, Counter Events and Async Events.

The Flow Events, Object Events, Metadata Events, Memory Dump Events, Mark Events and Clock Sync Event and Context Events are not supported.

Recording of C++ stack frames within events are not supported.

Added an implementation of a JSON Trace Event Logger that logs to the Google Trace Event Format.

Added UnitTest and Benchmarks test that verifies the implementation of the JSON Trace Event Logger.

Add implementation of a stack allocator for rapidjson to use to allocate memory from a pre-defined buffer.
This is used to have a `rapidjson::Writer` produce the JSON output for an event instead of hand-rolling a JSON solution.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>


## How was this PR tested?

Added UnitTest which outputs sample test data

```json
[
  {"name":"Duration Event","id":"0","cat":"Test","ph":"B","ts":1664329933375019,"pid":31760,"tid":36036,"args":{"string":"Hello world","int64_t":-2,"uint64_t":18446462603027808255,"bool":true,"double":64.0,"array":["Hello world",-2,18446462603027808255,true,64.0],"object":{"Field1":"Hello world","Field2":-2,"Field3":18446462603027808255,"Field4":true,"Field5":64.0}}},
  {"name":"Duration Event","id":"0","cat":"Test","ph":"E","ts":1664329933375119,"pid":31760,"tid":36036,"args":{"string":"Hello world","int64_t":-2,"uint64_t":18446462603027808255,"bool":true,"double":64.0,"array":["Hello world",-2,18446462603027808255,true,64.0],"object":{"Field1":"Hello world","Field2":-2,"Field3":18446462603027808255,"Field4":true,"Field5":64.0}}},
  {"name":"Complete Event","id":"0","cat":"Test","ph":"X","ts":1664329933375182,"pid":31760,"tid":36036,"args":{"string":"Hello world","int64_t":-2,"uint64_t":18446462603027808255,"bool":true,"double":64.0,"array":["Hello world",-2,18446462603027808255,true,64.0],"object":{"Field1":"Hello world","Field2":-2,"Field3":18446462603027808255,"Field4":true,"Field5":64.0}}},
  {"name":"Instant Event","id":"0","cat":"Test","ph":"i","ts":1664329933375235,"pid":31760,"tid":36036,"args":{"string":"Hello world","int64_t":-2,"uint64_t":18446462603027808255,"bool":true,"double":64.0,"array":["Hello world",-2,18446462603027808255,true,64.0],"object":{"Field1":"Hello world","Field2":-2,"Field3":18446462603027808255,"Field4":true,"Field5":64.0}}},
  {"name":"Duration Event","id":"1","cat":"Test","ph":"B","ts":1664329933375233,"pid":31760,"tid":1936,"args":{"string":"Hello world","int64_t":-2,"uint64_t":18446462603027808255,"bool":true,"double":64.0,"array":["Hello world",-2,18446462603027808255,true,64.0],"object":{"Field1":"Hello world","Field2":-2,"Field3":18446462603027808255,"Field4":true,"Field5":64.0}}},
  {"name":"Counter Event","id":"0","cat":"Test","ph":"C","ts":1664329933375286,"pid":31760,"tid":36036,"args":{"string":"Hello world","int64_t":-2,"uint64_t":18446462603027808255,"bool":true,"double":64.0,"array":["Hello world",-2,18446462603027808255,true,64.0],"object":{"Field1":"Hello world","Field2":-2,"Field3":18446462603027808255,"Field4":true,"Field5":64.0},"frameTime":16}},
  {"name":"Duration Event","id":"1","cat":"Test","ph":"E","ts":1664329933375353,"pid":31760,"tid":1936,"args":{"string":"Hello world","int64_t":-2,"uint64_t":18446462603027808255,"bool":true,"double":64.0,"array":["Hello world",-2,18446462603027808255,true,64.0],"object":{"Field1":"Hello world","Field2":-2,"Field3":18446462603027808255,"Field4":true,"Field5":64.0}}},
  {"name":"Duration Event","id":"2","cat":"Test","ph":"B","ts":1664329933375328,"pid":31760,"tid":33928,"args":{"string":"Hello world","int64_t":-2,"uint64_t":18446462603027808255,"bool":true,"double":64.0,"array":["Hello world",-2,18446462603027808255,true,64.0],"object":{"Field1":"Hello world","Field2":-2,"Field3":18446462603027808255,"Field4":true,"Field5":64.0}}},
  {"name":"Async Event","id":"0","cat":"Test","ph":"b","ts":1664329933375356,"pid":31760,"tid":36036,"args":{"string":"Hello world","int64_t":-2,"uint64_t":18446462603027808255,"bool":true,"double":64.0,"array":["Hello world",-2,18446462603027808255,true,64.0],"object":{"Field1":"Hello world","Field2":-2,"Field3":18446462603027808255,"Field4":true,"Field5":64.0}}},
  {"name":"Duration Event","id":"3","cat":"Test","ph":"B","ts":1664329933375355,"pid":31760,"tid":19832,"args":{"string":"Hello world","int64_t":-2,"uint64_t":18446462603027808255,"bool":true,"double":64.0,"array":["Hello world",-2,18446462603027808255,true,64.0],"object":{"Field1":"Hello world","Field2":-2,"Field3":18446462603027808255,"Field4":true,"Field5":64.0}}},
  {"name":"Complete Event","id":"1","cat":"Test","ph":"X","ts":1664329933375422,"pid":31760,"tid":1936,"args":{"string":"Hello world","int64_t":-2,"uint64_t":18446462603027808255,"bool":true,"double":64.0,"array":["Hello world",-2,18446462603027808255,true,64.0],"object":{"Field1":"Hello world","Field2":-2,"Field3":18446462603027808255,"Field4":true,"Field5":64.0}}},
  {"name":"Async Instant Event","id":"0","cat":"Test","ph":"n","ts":1664329933375436,"pid":31760,"tid":36036,"args":{"string":"Hello world","int64_t":-2,"uint64_t":18446462603027808255,"bool":true,"double":64.0,"array":["Hello world",-2,18446462603027808255,true,64.0],"object":{"Field1":"Hello world","Field2":-2,"Field3":18446462603027808255,"Field4":true,"Field5":64.0}}},
  {"name":"Duration Event","id":"2","cat":"Test","ph":"E","ts":1664329933375458,"pid":31760,"tid":33928,"args":{"string":"Hello world","int64_t":-2,"uint64_t":18446462603027808255,"bool":true,"double":64.0,"array":["Hello world",-2,18446462603027808255,true,64.0],"object":{"Field1":"Hello world","Field2":-2,"Field3":18446462603027808255,"Field4":true,"Field5":64.0}}},
  {"name":"Instant Event","id":"1","cat":"Test","ph":"i","ts":1664329933375475,"pid":31760,"tid":1936,"args":{"string":"Hello world","int64_t":-2,"uint64_t":18446462603027808255,"bool":true,"double":64.0,"array":["Hello world",-2,18446462603027808255,true,64.0],"object":{"Field1":"Hello world","Field2":-2,"Field3":18446462603027808255,"Field4":true,"Field5":64.0}}},
  {"name":"Duration Event","id":"3","cat":"Test","ph":"E","ts":1664329933375475,"pid":31760,"tid":19832,"args":{"string":"Hello world","int64_t":-2,"uint64_t":18446462603027808255,"bool":true,"double":64.0,"array":["Hello world",-2,18446462603027808255,true,64.0],"object":{"Field1":"Hello world","Field2":-2,"Field3":18446462603027808255,"Field4":true,"Field5":64.0}}},
  {"name":"Async Event","id":"0","cat":"Test","ph":"e","ts":1664329933375484,"pid":31760,"tid":36036,"args":{"string":"Hello world","int64_t":-2,"uint64_t":18446462603027808255,"bool":true,"double":64.0,"array":["Hello world",-2,18446462603027808255,true,64.0],"object":{"Field1":"Hello world","Field2":-2,"Field3":18446462603027808255,"Field4":true,"Field5":64.0}}},
  {"name":"Complete Event","id":"2","cat":"Test","ph":"X","ts":1664329933375506,"pid":31760,"tid":33928,"args":{"string":"Hello world","int64_t":-2,"uint64_t":18446462603027808255,"bool":true,"double":64.0,"array":["Hello world",-2,18446462603027808255,true,64.0],"object":{"Field1":"Hello world","Field2":-2,"Field3":18446462603027808255,"Field4":true,"Field5":64.0}}},
  {"name":"Counter Event","id":"1","cat":"Test","ph":"C","ts":1664329933375525,"pid":31760,"tid":1936,"args":{"string":"Hello world","int64_t":-2,"uint64_t":18446462603027808255,"bool":true,"double":64.0,"array":["Hello world",-2,18446462603027808255,true,64.0],"object":{"Field1":"Hello world","Field2":-2,"Field3":18446462603027808255,"Field4":true,"Field5":64.0},"frameTime":17}},
  {"name":"Complete Event","id":"3","cat":"Test","ph":"X","ts":1664329933375527,"pid":31760,"tid":19832,"args":{"string":"Hello world","int64_t":-2,"uint64_t":18446462603027808255,"bool":true,"double":64.0,"array":["Hello world",-2,18446462603027808255,true,64.0],"object":{"Field1":"Hello world","Field2":-2,"Field3":18446462603027808255,"Field4":true,"Field5":64.0}}},
  {"name":"Instant Event","id":"2","cat":"Test","ph":"i","ts":1664329933375554,"pid":31760,"tid":33928,"args":{"string":"Hello world","int64_t":-2,"uint64_t":18446462603027808255,"bool":true,"double":64.0,"array":["Hello world",-2,18446462603027808255,true,64.0],"object":{"Field1":"Hello world","Field2":-2,"Field3":18446462603027808255,"Field4":true,"Field5":64.0}}},
  {"name":"Async Event","id":"1","cat":"Test","ph":"b","ts":1664329933375573,"pid":31760,"tid":1936,"args":{"string":"Hello world","int64_t":-2,"uint64_t":18446462603027808255,"bool":true,"double":64.0,"array":["Hello world",-2,18446462603027808255,true,64.0],"object":{"Field1":"Hello world","Field2":-2,"Field3":18446462603027808255,"Field4":true,"Field5":64.0}}},
  {"name":"Instant Event","id":"3","cat":"Test","ph":"i","ts":1664329933375575,"pid":31760,"tid":19832,"args":{"string":"Hello world","int64_t":-2,"uint64_t":18446462603027808255,"bool":true,"double":64.0,"array":["Hello world",-2,18446462603027808255,true,64.0],"object":{"Field1":"Hello world","Field2":-2,"Field3":18446462603027808255,"Field4":true,"Field5":64.0}}},
  {"name":"Counter Event","id":"2","cat":"Test","ph":"C","ts":1664329933375605,"pid":31760,"tid":33928,"args":{"string":"Hello world","int64_t":-2,"uint64_t":18446462603027808255,"bool":true,"double":64.0,"array":["Hello world",-2,18446462603027808255,true,64.0],"object":{"Field1":"Hello world","Field2":-2,"Field3":18446462603027808255,"Field4":true,"Field5":64.0},"frameTime":18}},
  {"name":"Async Instant Event","id":"1","cat":"Test","ph":"n","ts":1664329933375625,"pid":31760,"tid":1936,"args":{"string":"Hello world","int64_t":-2,"uint64_t":18446462603027808255,"bool":true,"double":64.0,"array":["Hello world",-2,18446462603027808255,true,64.0],"object":{"Field1":"Hello world","Field2":-2,"Field3":18446462603027808255,"Field4":true,"Field5":64.0}}},
  {"name":"Counter Event","id":"3","cat":"Test","ph":"C","ts":1664329933375629,"pid":31760,"tid":19832,"args":{"string":"Hello world","int64_t":-2,"uint64_t":18446462603027808255,"bool":true,"double":64.0,"array":["Hello world",-2,18446462603027808255,true,64.0],"object":{"Field1":"Hello world","Field2":-2,"Field3":18446462603027808255,"Field4":true,"Field5":64.0},"frameTime":19}},
  {"name":"Async Event","id":"2","cat":"Test","ph":"b","ts":1664329933375654,"pid":31760,"tid":33928,"args":{"string":"Hello world","int64_t":-2,"uint64_t":18446462603027808255,"bool":true,"double":64.0,"array":["Hello world",-2,18446462603027808255,true,64.0],"object":{"Field1":"Hello world","Field2":-2,"Field3":18446462603027808255,"Field4":true,"Field5":64.0}}},
  {"name":"Async Event","id":"1","cat":"Test","ph":"e","ts":1664329933375675,"pid":31760,"tid":1936,"args":{"string":"Hello world","int64_t":-2,"uint64_t":18446462603027808255,"bool":true,"double":64.0,"array":["Hello world",-2,18446462603027808255,true,64.0],"object":{"Field1":"Hello world","Field2":-2,"Field3":18446462603027808255,"Field4":true,"Field5":64.0}}},
  {"name":"Async Event","id":"3","cat":"Test","ph":"b","ts":1664329933375677,"pid":31760,"tid":19832,"args":{"string":"Hello world","int64_t":-2,"uint64_t":18446462603027808255,"bool":true,"double":64.0,"array":["Hello world",-2,18446462603027808255,true,64.0],"object":{"Field1":"Hello world","Field2":-2,"Field3":18446462603027808255,"Field4":true,"Field5":64.0}}},
  {"name":"Async Instant Event","id":"2","cat":"Test","ph":"n","ts":1664329933375701,"pid":31760,"tid":33928,"args":{"string":"Hello world","int64_t":-2,"uint64_t":18446462603027808255,"bool":true,"double":64.0,"array":["Hello world",-2,18446462603027808255,true,64.0],"object":{"Field1":"Hello world","Field2":-2,"Field3":18446462603027808255,"Field4":true,"Field5":64.0}}},
  {"name":"Async Instant Event","id":"3","cat":"Test","ph":"n","ts":1664329933375724,"pid":31760,"tid":19832,"args":{"string":"Hello world","int64_t":-2,"uint64_t":18446462603027808255,"bool":true,"double":64.0,"array":["Hello world",-2,18446462603027808255,true,64.0],"object":{"Field1":"Hello world","Field2":-2,"Field3":18446462603027808255,"Field4":true,"Field5":64.0}}},
  {"name":"Async Event","id":"2","cat":"Test","ph":"e","ts":1664329933375751,"pid":31760,"tid":33928,"args":{"string":"Hello world","int64_t":-2,"uint64_t":18446462603027808255,"bool":true,"double":64.0,"array":["Hello world",-2,18446462603027808255,true,64.0],"object":{"Field1":"Hello world","Field2":-2,"Field3":18446462603027808255,"Field4":true,"Field5":64.0}}},
  {"name":"Async Event","id":"3","cat":"Test","ph":"e","ts":1664329933375770,"pid":31760,"tid":19832,"args":{"string":"Hello world","int64_t":-2,"uint64_t":18446462603027808255,"bool":true,"double":64.0,"array":["Hello world",-2,18446462603027808255,true,64.0],"object":{"Field1":"Hello world","Field2":-2,"Field3":18446462603027808255,"Field4":true,"Field5":64.0}}}
]
```

Added benchmark test to just test throughput of writing metrics to an in-memory stream or to a file directly.
```
Run on (64 X 3700 MHz CPU s)
CPU Caches:
  L1 Data 32K (x32)
  L1 Instruction 32K (x32)
  L2 Unified 524K (x32)
  L3 Unified 16777K (x8)
------------------------------------------------------------------------------------------------------------------------------------------------------
Benchmark                                                                                                            Time             CPU   Iterations
------------------------------------------------------------------------------------------------------------------------------------------------------
JsonTraceEventLoggerBenchmarkFixture/BM_JsonTraceEventLogger_RecordEventsToInMemoryStream                        30906 ns        30030 ns        21333
JsonTraceEventLoggerBenchmarkFixture/BM_JsonTraceEventLogger_RecordEventsToTemporaryDirUnbufferedFileStream      54867 ns        57199 ns        11200
OKAY AzRunBenchmarks() returned 0
Returning code: 0```